Feat: 인터넷 연결 없을 시 화면
