+++ package.json
@@ -10,6 +10,7 @@
     "test": "jest"
   },
   "dependencies": {
+    "@react-native-community/netinfo": "^11.3.2",
     "react": "18.2.0",
     "react-native": "0.74.3",
     "react-native-webview": "^13.10.5"
